fixup! Add test automation script that tests the rootfs requirements
do not call functions in context of if etc

The if statement has an (intended) side effect that the exit status of the
argument does not cause set -e cause an exit.

For example, if I would do
if grep <string> <file>; then
    found
else
    notfound
fi

then grep will return either 0 or 1 depending if it found the string.
If it returns 1, i do not want entire shell script to exit of course.

if you use
if setup; then ...

anything run in the setup function, is temporally run with set -e
disabled. As such, none of the commands cause an exit and thus during
failure anything in setup keeps going on happily.