buildenv: search for tools also in the sbin paths
Some tools require sudo permissions to run, and thus they are not in
the users path by default. Extend the path to include the common sbin
locations to allow this script to run without being sudo. We only need
to know if we can execute them, not if we have all permissions to
perform all tasks of the tool.

Addresses issue EMP-271

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>