build: bind mount qemu-arm
To be able to run arm binaries, in our case during the image test, we
need to be able to run them using an arm binary interpreter, aka
qemu-arm. As it depends on who registers the binary in
/proc/sys/fs/binfmt_misc/register it is impossible to know or assume
which interpreter to use inside of the build environment.

For example the host may have installed qemu-arm-static, but the
container tries to run qemu-arm.

The kernel provides a method to obtain the current interpreter for a
given signature. To avoid having to use the same routine everywhere and
ensuring it is mounted everywhere, the ARM_EMU_BIN variable and path is
passed into each layer. It is up to the host to ensure this is passed
along nicely. For local builds, the build_for_ultimaker.sh script takes
care of this. For the CI, the setup there is responsible.
When running the individual scripts, the caller is responsible to setup
things accordingly and truthfully.

Addresses to issue EMP-271 and fixes issue EMP-356

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>