tests/toolbox_image: do not capture the return code just to return it
What we currently have:

( <cmd> && return 0 ) || return 1

In other words, run <cmd> if it ran successfully by returning 0 run the
following command, return 0. If either of these failed, return 1.

Which may have its uses but even that could be simplified to

<cmd> || return 1

which would be useful if we always want to return 1 on error, no matter
what <cmd> returns.

So just return whatever <cmd> returns.

Noted is that chroot returns whatever <cmd> passed to chroot returns,
so the chroot exit code is acceptable.

Finally, to argue, that we are not returning anything when not using
return is not true.

some_function_init_stuff
call afunction
return $?

In other words, the last commands exit code is always 'returned' anyway.

Addresses EMP-271

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>