buildenv: Add qemu to the buildenv test CI job
To be able to run arm emulated code, two things are required. The
QEMU_EMU_BIN variable needs to be populated with the path to the local
arm emulator, and this binary needs to be mounted into the container.

The build_for_ultimaker.sh script already does this, the CI did not.

This is required to fix the dockerfile from failing to build.

Addresses issue EMP-272, fixes EMP-371.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>