Format partitions to honor the new disk size.
Add a function that checks the partition layout and resizes partitions
where needed or reformats them if resize is not possible.

This pulls in blkid as we need to identify the partition type to be
able to call the correct tools.

This script does make the assumption that all partitions are f2fs
based, except for the BOOT partition, which is ext4 until u-boot
supports f2fs.

Addresses EMP-272, closes EMP-378

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>