scripts/prepare_disk: force formatting of partition
The mkfs.f2fs program sometimes fails to 'detect' the filesystem type
on a partition when the primary superblock is corrupted, but not
entirely and will refuse to run unless it is forced.

Since the script already made the decision to to format, so force it to
ensure the script can continue and not fail heron.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>