local-build: Always pull build environment in docker
When using docker run, it will only pull the latest image if it does
not exist yet. To ensure we always have the latest image, we need to
run docker pull. However that interferes with the development approach,
where we need a newer (modified) image, that we haven't pushed yet (and
shouldn't).

This patch first tries to pull the image, and if that fails, builds it
locally instead using the supplied tag name.

Address EMP-272, closes EMP-393.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>