Add update rootfs files script
This script takes two argument, the 'update source' and the 'target
storage device' and synchronizes (source -> target) the files to
the second partitition of the given target device. The arguments
can be passed as a parameter or through the environment. Excluding
files or directories can be done through the exlude list.

Contributes to EMP-273, adresses EMP-311

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>