test/buildenv: Check for execute ability of arm emu
The build environment checks for an arm emulator with various tests,
but never tries to actually execute the emulator. By not doing this,
we can miss the fact that ARM_EMU_BIN exists as executable, but fails
to run due to missing shared libraries for example.

This patch adds a simple test to see whether the binary can actually
run.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>