build: Store executable state in git
We should use our version control system to store the executable state
of a binary/script in git. As such, we should also not force the
executable state during the build step, but actually only run the
script if it is marked as executable on the target.

Attributes to EMP-382, contributes to EMP-416

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>