scripts/start_update: Do not exec functions in an if
Trapping a function in an if statement, causes set -e and set -u to no
longer function. This is by design. This would have as a consequence
that every command now needs to be appended with || return 1.

This is error prone when return statements are missing for example, steps would
still continue even though something is wrong.

Instead, we have to ensure that our cleanup steps and calling function
do the proper thing on failure.

Fail early, fail often.

Addresses to EMP-382, contributes to EMP-416

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>