scripts/start_update: Do not use an unneeded temporary dir
Currently, we create a temporary directory, for example /tmp/tempdir1
and then bind mount this to /toolbox_mount/update_source.

We do this because /toolbox_mount/update_source is not writable. Much
easier is of course to mount a tmpfs directory onto this directory
making the it writable instead.

Contributes to EMP-416

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>