tar: change how arguments are supplied
Command line utilities tend to very often use hyphens. We also use
hyphens throughout the project, except on some instances of tar. Let us
also use a hyphen there to be consistent.

Further more, we should synchronize the order. Most importantly, the f
parameter, indicating a file is to come, should be last, or rather
in front of the file.

Contributes to EMP-416

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>