filters: Split excludes into files
Allow the user of both .keep and .discard files, effectively splitting
the 'exclude' file. This allow for more fine grained control of what to
keep/remove, for example per application a single keep file (for now
the file is stored in this repository).

Also by splitting is needed, that when saving files, we are only
interested in the '.keep' part, vs when rsyncing, we care about both.

In other words, '/tmp' and /media' are not something that needs to be
saved, nor do we want to rsync it, it should be ignored during jobs.

To satisfy several tools, the paths have been changed to relative paths vs
absolute paths.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>